Rename scan_path_controller to controller
